Solved the issue in the strong wash-out regime and few minor changes
Substituted the diagonal matrix of the lepton chemical potentials with a vector in the equations for the lepton asymmetry. This solved the instabilities arising for some sets of parameter points with large ys in the strong wash-out regime. Thanks Brian for finding this insidious typo and for providing the solution to the problem.

I have also changed the output to etaB. The code was returning |etaB|, but it is important to know the sign of the BAU for a given parameter point.